+++ chicken_dinner/models/telemetry/events.py
@@ -17,7 +17,8 @@ class TelemetryEvent(object):
     embedded telemetry objects recursively. Converts all event and object keys
     to snake-cased key names.
 
-    :param data: the JSON object data associated with the telemetry event
+    :param dict data: the JSON object data associated with the telemetry event
+    :param bool map_assets: whether to map asset ids to asset names
     """
 
     def __init__(self, data, map_assets=False):
@@ -71,3 +72,18 @@ class TelemetryEvent(object):
     def to_dict(self):
         """Get the event as a dict."""
         return json.loads(self.dumps())
+
+    def keys(self):
+        """Get all attributes names."""
+        return [k for k in self.__dict__.keys() if k[0] != "_"]
+
+    def values(self):
+        """Get all attribute values."""
+        return [self.__dict__[k] for k in self.__dict__.keys() if k[0] != "_"]
+
+    def items(self):
+        """Iterate through the attributes dictionary."""
+        for k, v in self.__dict__.items():
+            if k[0] == "_":
+                continue
+            yield k, v